requirements.txt
streamlit==1.38.0
reportlab==4.2.2
python-docx==1.1.2
PyYAML==6.0.2
